+++ src/App.js
@@ -1,18 +1,18 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Redirect, Switch } from 'react-router-dom';
 import './App.css';
-import axios from 'axios';
+
 
 import AdminMainPage from './adminComponent/AdminMainPage'
-import StudentPage from './Calendar/StudentPage.jsx';
-import StaffPage from './Calendar/StaffPage.jsx';
+import StudentPage from './Calendar/StudentComponent/StudentPage.jsx';
+import StaffPage from './Calendar/StaffComponent/StaffPage.jsx';
 import CalendarStore from './mobx/CalendarStore.js';
-import ContentRouting from './components/contentRouting/contentRouting.jsx';
+import ContentRouting from './TaskViewComponent/StudentTaskView/contentRouting.jsx';
 import {AdminOnlyRoute, StudentOnlyRoute, StaffOnlyRoute, PrivateRoute, LoginRoute, CheckSwitchRoute, RegisterRoute} from './SpecialRoutes';
-import ProjectListPage from './components/projectListing.jsx';
-import LoginPage from "./components/LoginPage";
-import RegisterPage from "./components/RegisterUser";
-import Kanban from "./components/contentRouting/TaskBoard";
+import ProjectListPage from './TaskViewComponent/StaffTaskView/projectListing.jsx';
+import LoginPage from "./LoginComponent/LoginPage";
+import RegisterPage from "./LoginComponent/RegisterUser";
+
