+++ src/Calendar/ReusableCalendar.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { observer } from "mobx-react";
-import StudentCalendar from "./StudentCalendar.jsx";
-import StaffCalendar from "./StaffCalendar.jsx";
+import StudentCalendar from "./StudentComponent/StudentCalendar.jsx";
+import StaffCalendar from "./StaffComponent/StaffCalendar.jsx";
 import { withRouter } from "react-router-dom";
