+++ src/TaskViewComponent/StudentTaskView/contentRouting.jsx
@@ -16,7 +16,6 @@ import clsx from 'clsx';
 import * as actions from '../../redux/login-store/actions/authActions';
 
 import MeetingContentPage from './MeetingContentPage';
-import ReportContentPage from './ReportContentPage';
 import TaskBoardPage from './TaskBoardPage';
 import SubmissionContentPage from './SubmissionContentPage';
 import { connect } from 'react-redux';
